Add a 4th tab that demos some attachment behaviors.
Allow dragging of a view via an attachment and pan gesture.
Attach two items with a non-rigid spring-like attachment.

Add a custom dynamic item that can be used to log dynamic item changes for debugging.

Make the force of the push behavior for buttons scale with the area of the view.